requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-telegram-bot[job-queue]==21.5
psycopg2-binary==2.9.9
pydantic==2.8.2
